Unifying System.Reflection.Extensions and Uap10.1 packaging for System.Reflection
Remove dead code from System.Reflection.Primitives

LowLevelEnumerable is not used anymore. Removing it
