+++ src/System.Reflection.Extensions/pkg/System.Reflection.Extensions.builds
@@ -3,8 +3,6 @@
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
   <ItemGroup>
     <Project Include="System.Reflection.Extensions.pkgproj" />
-    <Project Include="any\System.Reflection.Extensions.pkgproj" />
-    <Project Include="aot\System.Reflection.Extensions.pkgproj" />
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.traversal.targets))\dir.traversal.targets" />
 </Project>
