--- src/System.Reflection/src/netcore50aot/project.json
@@ -1,9 +1,0 @@
-{
-  "frameworks": {
-    "netcore50": {
-      "dependencies": {
-        "Microsoft.TargetingPack.Private.NETNative": "1.1.0-beta-24520-00"
-      }
-    }
-  }
-}
