+++ src/System.Reflection/src/project.json
@@ -2,7 +2,7 @@
   "frameworks": {
     "netstandard1.7": {
       "dependencies": {
-        "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24520-03"
+        "System.Runtime": "4.3.0-beta-24516-03"
       },
       "imports": [
         "dotnet5.8"
@@ -11,6 +11,11 @@
     "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
+      }
+    },
+    "uap10.1": {
+      "dependencies": {
+        "System.Runtime": "4.3.0-beta-24516-03"
       }
     }
   }
