+++ src/System.Xml.XmlSerializer/src/project.json
@@ -43,7 +43,7 @@
     },
     "uap10.1": {
       "dependencies": {
-        "Microsoft.TargetingPack.Private.NETNative": "1.1.0-beta-24520-00",
+        "Microsoft.TargetingPack.Private.NETNative": "1.1.0-beta-24520-01",
         "System.Collections.NonGeneric": "4.3.0-beta-24520-06",
         "System.Runtime": "4.3.0-beta-24520-06"
       }
